Added rollout.js. Moved getConfig and checkConfig into authorize.
